Swalbe

Swalbe.SysConst

Swalbe.BGKandStream!

Swalbe.Sys

Swalbe.boxpattern

Swalbe.ellipsepattern

Swalbe.equilibrium!

Swalbe.filmpressure!

Swalbe.moments!

Swalbe.power_broad

Swalbe.randinterface!

Swalbe.run_dropletforced

Swalbe.run_dropletpatterned

Swalbe.run_dropletrelax

Swalbe.run_flat

Swalbe.run_random

Swalbe.run_rayleightaylor

Swalbe.sinewave1D!

Swalbe.singledroplet

Swalbe.slippage!

Swalbe.thermal!

Swalbe.trianglepattern

Swalbe.viewdists

Swalbe.viewneighbors

Swalbe.∇f!

Swalbe.∇²f!

Swalbe.SysConst — Type

SysConst{T}

Struct that contains all run time constants, e.g. lattice size, surface tension γ and so on.

Swalbe.BGKandStream! — Method

BGKandStream!(fout, feq, ftemp, Fx, Fy, τ)

Performs a BGK collision operation with a WFM forcecorrection and a subsequent streaming of the resulting populations.

Arguments

fout :: Array{<:Number,3}: Streamed distribution after the collision processes

feq :: Array{<:Number,3}: Equilibrium distribution, computed with equilibrium!

ftemp :: Array{<:Number,3}: Temporary distribution from the time step before, only useful if $\tau \neq 1$

Fx :: Array{<:Number,2}: Sum of forces acting on the fluid in x-direction

Fy :: Array{<:Number,2}: Sum of forces acting on the fluid in y-direction

τ <: Number: Relaxtion time, if not supplied $\tau = 1$ assumed

Mathematics

The lattice Boltzmann equation in its discretized format is relatively simple to write down

$f_{\alpha}(\mathbf{x}+\mathbf{e}_{\alpha}\Delta t, t+\Delta t) - f_{\alpha}(\mathbf{x}, t) = -\frac{1}{\tau}(f_{\alpha}(\mathbf{x}, t) - f^{\text{eq}}_{\alpha}(\mathbf{x}, t)) + \Delta t \mathcal{S}_{\alpha},$

where the collision kernel is approximated with a BKG single relaxation time (SRT)

$\Omega_{\alpha} = \frac{1}{\tau}(f_{\alpha}(\mathbf{x}, t) - f^{\text{eq}}_{\alpha}(\mathbf{x}, t),$

and a source term $\mathcal{S}$ which is given by

$\mathcal{S}_{\alpha} = \frac{3 w_{\alpha}}{e_{\alpha x}e_{\alpha x}+e_{\alpha y}e_{\alpha y}}\mathbf{e_{\alpha}}\cdot\mathbf{F}_{\alpha} .$

The term $e_{\alpha x}e_{\alpha x}+e_{\alpha y}e_{\alpha y}$ is either zero for the zeroth population, 1 for the first four populations or 2 for the remaining ones.

Examples

julia> using Swalbe

julia> feq = ones(5,5,9); ftemp = zeros(5,5,9); fout = zeros(5,5,9);

julia> feq[1,1,:] .= 2.0 # To check the streaming process 
9-element view(::Array{Float64,3}, 1, 1, :) with eltype Float64:
 2.0
 2.0
 2.0
 2.0
 2.0
 2.0
 2.0
 2.0
 2.0

julia> Swalbe.BGKandStream!(fout, feq, ftemp, zeros(5,5), zeros(5,5))

julia> fout[:,:,6] # The value 2 should have moved one down and one to the right!
5×5 Array{Float64,2}:
 1.0  1.0  1.0  1.0  1.0
 1.0  2.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0  1.0
 1.0  1.0  1.0  1.0  1.0

References

Salmon

Dellar

Peng et al.

See also: Swalbe.equilibrium

Swalbe.Sys — Method

Sys(sysc, device, exotic)

Mostly allocations of arrays used to run a simulation, but all within one function :)

Swalbe.boxpattern — Method

boxpattern(θ, θ₀; center=(size(θ,1)÷2, size(θ,2)÷2), δₐ=1/36, side=20)

Defines a quadratic box around a center with side length side.

Arguments

θ::Array{undef,2}: Contact angle map

θ₀: Default contact angle θ

center::Tuple{Int, Int}: center of the box pattern, default value is (size(θ,1)÷2, size(θ,2)÷2)

δₐ::Float64: Contact angle contrast with the substrate, default is set to 1/36 ≈ 5 degrees difference

side::Int: Length of the sides, default is set to 20

Examples

julia> using Swalbe

julia> θ₀ = 1/9;

julia> pattern, polygon = Swalbe.boxpattern(ones(100,100), θ₀);

julia> polygon # Some cool thing we use to create the posize(θ,2)gones, a LazySet
LazySets.VPolygon{Float64,Array{Float64,1}}([[40.0, 40.0], [60.0, 40.0], [60.0, 60.0], [40.0, 60.0]])

julia> pattern[50,50] == pattern[1,1] # In the center there is a different contact angle!
false

References

Really not much to say here, check out LazySets.jl.

Swalbe.ellipsepattern — Method

ellipsepattern(θ, θ₀; center=(size(θ,1)÷2, size(θ,2)÷2), δₐ=1/36, a=10, b=5)

Creates an ellipse shaped contact angle defect with contact angle mismatch δₐ.

Arguments

θ::Array{undef,2}: Contact angle map

θ₀: Default contact angle θ

center::Tuple{Int, Int}: center of the created pattern, default values is center = (size(θ,1)÷2, size(θ,2)÷2)

δₐ::Float64: contact angle mismatch between patch and rest of substrate, default is δₐ = 1/36 or 5 degrees

a::Int:: semimajor half ax of the ellipse, default value is a=10

b::Int:: semiminor half ax of the ellipse, default value is b=5

Examples

julia> using Swalbe, Test

julia> θ₀ = 1/9;

julia> θ, P = Swalbe.ellipsepattern(ones(100,100), θ₀); # per default the center is in the middle!

julia> @test θ[1,1] == θ₀
Test Passed

julia> @test θ[50,50] == θ₀ + 1/36 # The default increment, is about 5 degrees.
Test Passed

References

Nothing interesting here.

Swalbe.equilibrium! — Method

Equilibrium!(feq, height,velocityx, velocityy, vsquare, gravity)

Calculation of the equilibrium distribution feq for the shallow water lattice Boltzmann method.

Arguments

feq :: Array{<:Number,3}: Equilibrium distribution function, to be calculated

height :: Array{<:Number,2}: The height field $h(\mathbf{x},t)$

velocityx :: Array{<:Number,2}: x-component of the macroscopic velocity

velocityy :: Array{<:Number,2}: y-component of the macroscopic velocity

vsquare :: Array{<:Number,2}: Dummy array that is preallocated to be filled with the square of the velocity vector

gravity <: Number: Strength of the gravitational acceleration in lattice units

Mathematics

The detailed motivation and derivation can be found in the article of Salmon. Similar to the standard Navier-Stokes approximating lattice Boltzmann methods the equilibrium distribution feq is an expansion to second order in velocity u. If you want it is an slow mode part of the shallow water theory and thus the equilibrium is given as

$f_i^{\text{eq}} = h \bigg(1 - \frac{5}{6}g h - \frac{2}{3}u^2\bigg),\quad i=0 \newline f_i^{\text{eq}} = w_i h \bigg(g h + 3 \mathbf{c}_i\cdot\mathbf{u} + \frac{9}{2}(\mathbf{c}_i\cdot\mathbf{u})^2) + \frac{3}{2}u^2\bigg),\quad else$

where $g$ is the gravitational acceleration (in lattice units) and $w_i, \mathbf{c}_i$ are the weights and lattice velocities.

It has been shown that it is possible to get rid of the gravity driven term in the equilibrium distribution, thus

$f_i^{\text{eq}} = h \bigg(1 - \frac{2}{3}u^2\bigg),\quad i=0 \newline f_i^{\text{eq}} = w_i h \bigg(3 \mathbf{c}_i\cdot\mathbf{u} + \frac{9}{2}(\mathbf{c}_i\cdot\mathbf{u})^2) + \frac{3}{2}u^2\bigg),\quad else$

then of course the topography gradient has to be included as a force term.

Examples

julia> using Swalbe, Test

julia> feq = zeros(5,5,9); ρ = ones(5,5); ux = fill(0.1,5,5); uy = zeros(5,5);

julia> Swalbe.equilibrium!(feq, ρ, ux, uy, zeros(5,5), 0.1) # Supply dummy u^2 as well.

julia> feq[:,:,1]
5×5 Array{Float64,2}:
 0.91  0.91  0.91  0.91  0.91
 0.91  0.91  0.91  0.91  0.91
 0.91  0.91  0.91  0.91  0.91
 0.91  0.91  0.91  0.91  0.91
 0.91  0.91  0.91  0.91  0.91

julia> Swalbe.equilibrium!(feq, ρ, ux, uy, zeros(5,5)) # Use the dispatch without g.

julia> @test all(feq[:,:,1] .≈ 1 - 2/3 * 0.01)
Test Passed

References

Salmon

Dellar

Peng et al.

Swalbe.filmpressure! — Method

filmpressure!(pressure, height, γ, θ, n, m, hmin, hcrit)

Calculation of the capillary pressure which is given by $p = - γ∇²h+ Π(h)$.

Arguments

pressure :: Array{Number,2}: Array that store the result of the compuation

height :: Array{Number,2}: Height field $h(\mathbf{x},t)$

γ <: Number: Forcing strenght due to surface tension

θ <: Number: Equilibrium contact angle

n :: Int: Larger power law exponent for $Π(h)$

m :: Int: Smaller power law exponent for $Π(h)$

hmin <: Number: Parameter of $Π(h)$, in fact $Π(hmin) = 0$

hcrit <: Number: Numerical stabilizer for case $h(\mathbf{x},t) \ll hmin$

Mathematics

The capillary pressure $p_{\text{cap}}$ is the centeral angle to match our model with the thin film equation. It consists of two parts, first being the laplace pressure $\nabla^2 h$ and second being the derivative of the disjoining pontential $\Pi(h)$

$p_{\text{cap}} = -\gamma \nabla^2 h + \Pi(h).$

For the laplacian term we use the same nine point discretization as in Swlabe.∇²f!. $\Pi(h)$ on the other hand is given by

$\Pi(h) = \kappa(\theta)f(h),$

where $\kappa(\theta)$ is simply a measure for the Hamaker constant and given as

$\kappa(\theta) = \gamma(1- \cos(\theta))\frac{(n-1)(m-1)}{(n-m)h_{\text{min}}}.$

For $f(h)$ one can use various forms, a very common however is the power law given by

$f(h) = \bigg[\bigg(\frac{h_{\text{min}}}{h}\bigg)^n - \bigg(\frac{h_{\text{min}}}{h}\bigg)^m\bigg].$

Examples

julia> using Swalbe, Test

julia> h = reshape(collect(1.0:25.0),5,5) # A dummy height field
5×5 Array{Float64,2}:
 1.0   6.0  11.0  16.0  21.0
 2.0   7.0  12.0  17.0  22.0
 3.0   8.0  13.0  18.0  23.0
 4.0   9.0  14.0  19.0  24.0
 5.0  10.0  15.0  20.0  25.0

julia> pressure = zeros(5,5); θ = 0.0; # Fully wetting substrate

julia> Swalbe.filmpressure!(pressure, h, θ) # default γ = 0.01

julia> result = [30.0 5.0 5.0 5.0 -20;
                 25.0 0.0 0.0 0.0 -25.0;
                 25.0 0.0 0.0 0.0 -25.0;
                 25.0 0.0 0.0 0.0 -25.0;
                 20.0 -5.0 -5.0 -5.0 -30.0];

julia> for i in eachindex(result)
           @test result[i] .≈ -100 .* pressure[i] atol=1e-12
       end

References

Peschka et al.

Craster and Matar

Derjaguin and Churaev

Swalbe.moments! — Method

moments!(height, velx, vely, fout)

Computation of the hydrodynamic moments, height and velocity.

Mathematics

The macroscopic quantities such as the height and the velocity are the moments of the distribution function,

$h(\mathbf{x},t) = \sum_{i=0}^8 f_i ,$

and

$\mathbf{v}(\mathbf{x},t) = \frac{1}{h}\sum_{i=0}^8 \mathbf{c}_i f_i .$

Examples

julia> using Swalbe, Test

julia> fout = zeros(5,5,9); fout[:,:,1] .= 1.0; fout[:,:,2] .= 0.1; # Dist with artifical velocity in x

julia> height = zeros(5,5); velx = zeros(5,5); vely = zeros(5,5);

julia> Swalbe.moments!(height,velx,vely,fout)

julia> @test all(height .== 1.1)
Test Passed

julia> @test all(velx .== 0.1/1.1)
Test Passed

julia> @test all(vely .== 0.0)
Test Passed

References

Krüger

Salmon

Zitz, Scagliarini and Harting

Swalbe.power_broad — Method

power_broad(arg, n)

Computes arg to the power n.

Actually this is useful because the ^ operator is much slower. Same thing I learned about the pow function in C, * yes it does what you want, but it is slow as fuck *.

Examples

julia> using Swalbe, Test

julia> Swalbe.power_broad(3, 3)
27

julia> Swalbe.power_broad.([2.0 5.0 6.0], 2) # Use the broadcasting operator `.`
1×3 Array{Float64,2}:
 4.0  25.0  36.0

See also: filmpressure

Swalbe.randinterface! — Method

randinterface!(height, h₀, ϵ)

Creates a random height field with average height h₀ and displacement magnitude ϵ.

Swalbe.run_dropletforced — Method

run_dropletforced()

Simulates an droplet on a patterned substrate

Swalbe.run_dropletpatterned — Method

run_dropletpatterned()

Simulates an droplet on a patterned substrate

Swalbe.run_dropletrelax — Method

run_dropletrelax()

Simulates an out of equilibrium droplet

Swalbe.run_flat — Method

run_flat(Sys::SysConst, device::String)

Performs a simulation of an flat interface without forces

Swalbe.run_random — Method

run_random(sys::SysConst, device::String)

Simulation of an random undulated interface

Swalbe.run_rayleightaylor — Method

run_rayleightaylor(sys::SysConst, device::String)

Simulation of an random undulated interface

Swalbe.sinewave1D! — Method

sinewave1D!(height, h₀, n, ϵ)

Creates a sine wave like height field with n full waves, average height h₀ and displacement magnitude ϵ.

Swalbe.singledroplet — Method

singledroplet(T, size(θ,1), size(θ,2), radius, θ, center, device=false)

Generates a fluid configuration of a single droplet in the shape of spherical cap with contact angle θ, sphere radius radius and centered at center.

Arguments

height::Array{undef, 2}: numerical formate, either Float64 or Float32

radius::AbstractFloat: radius of the underlying sphere from which the spherical cap is cut off

θ::AbstractFloat: contact angle in multiples of π

center::Tuple{Int,Int}: x and y coordinates of the center of the droplet

Examples

julia> using Swalbe, Test

julia> rad = 50; θ = 1/3;

julia> height = Swalbe.singledroplet(ones(100,100), rad, θ, (50,50));

julia> @test maximum(height) == rad * (1 - cospi(θ)) # Simple geometry
Test Passed

julia> argmax(height) # Which is constistent with the center!
CartesianIndex(50, 50)

References

See also:

Swalbe.slippage! — Method

slippage!(slipx, slipy, height, velx, vely, δ, μ)

Fluid substrate interaction that effectively mimics a velocity boundary condition at $h=0$.

Arguments

slipx :: Array{<:Number,2}: The x-component of the force due the velocity boundary condition

slipy :: Array{<:Number,2}: The y-component of the force due the velocity boundary condition

height::Array{<:Number,2}: Height field $h(\mathbf{x},t)$

velx::Array{<:Number,2}: x-component of the macroscopic velocity vector

vely::Array{<:Number,2}: y-component of the macroscopic velocity vector

δ <: Number: Extrapolation length into the substrate where the no-slip is met

μ <: Number: Kinematic viscosity of the simulation, dependent on the value of τ

Mathematics

With the velocity boundary condition at the fluid substrate interface we build the second main descriptor between our model and the thin film equation. One well studied assumption is that the fluid velocity vanishes at $h(\mathbf{x}) = 0$ which is called no slip condition.

TBD

Examples

TBD

References

Zitz, Scagliarini and Harting

Münch, Wagner und Witelski

Oron, Davis and Bankoff

Swalbe.thermal! — Method

thermal!(fx, fy, height, kᵦT, μ, δ)

Computations of force due to thermal fluctuations.

Arguments

fx :: Array{<:Number,2}: x-component of the force due to the fluctuations

fy :: Array{<:Number,2}: y-component of the force due to the fluctuations

height :: Array{<:Number,2}: Height field $h(\mathbf{x},t)$

kᵦT <: Number: Strenght of thermal fluctuations in lattice units

μ <: Number: The kinetic viscosity

δ <: Number: Slip length, needed to normalize

Mathematics

TBD

Examples

julia> using Swalbe, Statistics, Test

julia> x = ones(50,50); y = ones(50,50); h = ones(50,50);

julia> Swalbe.thermal!(x, y, h, 0.1, 1/6, 1)

julia> @test mean(x) ≈ 0.0 atol=1e-2
Test Passed

julia> @test mean(y) ≈ 0.0 atol=1e-2
Test Passed

julia> @test var(x) ≈ 2*0.1/11 atol=(2*0.1/11)/10 # var = 2kbt*6*μ/slip
Test Passed

References

Grün, Mecke and Rauscher

Mecke, Rauscher

Davidovitch, Moro and Stone

Swalbe.trianglepattern — Method

trianglepattern(θ, θ₀; center=(size(θ,1)÷2, size(θ,2)÷2), δₐ=1/36, side=60)

Generates an equilateral triangle centered around center with contact angle contrast δₐ and side length side.

Arguments

θ::Array{undef,2}: Contact angle map

θ₀: Default contact angle θ

center::Tuple(Int, Int): position of the center of the triangle

δₐ::Float64: contact angle contrast with the rest of the substrate

side::Int: length of the sides of the equilateral triangle

Examples

julia> using Swalbe

julia> θ, P = Swalbe.trianglepattern(ones(50,50), 1/9, side=20) # Returns a polygon and the contact angle field
([0.1111111111111111 0.1111111111111111 … 0.1111111111111111 0.1111111111111111; 0.1111111111111111 0.1111111111111111 … 0.1111111111111111 0.1111111111111111; … ; 0.1111111111111111 0.1111111111111111 … 0.1111111111111111 0.1111111111111111; 0.1111111111111111 0.1111111111111111 … 0.1111111111111111 0.1111111111111111], LazySets.VPolygon{Float64,Array{Float64,1}}([[15.0, 19.226497308103742], [35.0, 19.226497308103742], [25.0, 36.547005383792516]]))

julia> P
LazySets.VPolygon{Float64,Array{Float64,1}}([[15.0, 19.226497308103742], [35.0, 19.226497308103742], [25.0, 36.547005383792516]])

julia> θ[25,25]
0.1388888888888889

References

See also:

Swalbe.viewdists — Method

viewdists(f)

Generates a view for all nine populations of a D2Q9 distribution function.

Examples

julia> using Swalbe, Test

julia> ftest = reshape(collect(1.0:225.0),5,5,9);

julia> f0, f1, f2, f3, f4, f5, f6, f7, f8 = Swalbe.viewdists(ftest);

julia> @test all(f3 .== ftest[:,:,4])
Test Passed

See also: Swalbe.BGKandStream

Swalbe.viewneighbors — Method

viewneighbors(f)

Generates a view for all nine populations of a D2Q9 distribution function.

Examples

julia> using Swalbe, Test

julia> ftest = reshape(collect(1.0:5*5*8),5,5,8);

julia> f1, f2, f3, f4, f5, f6, f7, f8 = Swalbe.viewneighbors(ftest);

julia> @test all(f3 .== ftest[:,:,3])
Test Passed

See also: Swalbe.∇f!, Swalbe.filmpressure!

Swalbe.∇f! — Method

∇f!(outputx, outputy, f)

Gradient calculation with finite differences.

Computes both spatial first derivatives with a nine point stencil from an input f and writes the result to outputx and outputy. Since broadcasting is simple on the GPU we make use of circshift for the neighbors.

Mathematics

The gardient in two dimensions is given as

$\nabla f = \big(\frac{\partial f}{\partial x}, \frac{\partial f}{\partial y}\big)^T .$

Again with the nine point stencil this reduces to

$\frac{\partial f}{\partial x} = \frac{1}{3} (f_{i+1,j} - f_{i-1,j}) + \frac{1}{12}(f_{i+1,j+1} - f_{i-1,j+1} - f_{i-1,j-1} + f_{i+1,j-1}) ,$

and for the y component we get

$\frac{\partial f}{\partial y} = \frac{1}{3} (f_{i,j+1} - f_{i,j-1}) + \frac{1}{12}(f_{i+1,j+1} + f_{i-1,j+1} - f_{i-1,j-1} - f_{i+1,j-1}) .$

For the exact derivation feel free to read the reference by Junk and Klar.

Examples

julia> using Swalbe, Test

julia> arg = reshape(collect(1.0:25),5,5)
5×5 Array{Float64,2}:
 1.0   6.0  11.0  16.0  21.0
 2.0   7.0  12.0  17.0  22.0
 3.0   8.0  13.0  18.0  23.0
 4.0   9.0  14.0  19.0  24.0
 5.0  10.0  15.0  20.0  25.0

julia> resx = zeros(5,5); resy = zeros(5,5); Swalbe.∇f!(resx, resy, arg)

julia> whatXshouldbe = [-1.5 -1.5 -1.5 -1.5 -1.5;
                         1.0 1.0 1.0 1.0 1.0;
                         1.0 1.0 1.0 1.0 1.0;
                         1.0 1.0 1.0 1.0 1.0;
                        -1.5 -1.5 -1.5 -1.5 -1.5];

julia> for i in eachindex(resx) # Test the x-component
           @test resx[i] ≈ whatXshouldbe[i] atol=1e-10
       end

julia> whatYshouldbe = [-7.5 5.0 5.0 5.0 -7.5;
                        -7.5 5.0 5.0 5.0 -7.5;
                        -7.5 5.0 5.0 5.0 -7.5;
                        -7.5 5.0 5.0 5.0 -7.5;
                        -7.5 5.0 5.0 5.0 -7.5];

julia> for i in eachindex(resy) # Test the y-component
           @test resy[i] ≈ whatYshouldbe[i] atol=1e-10
       end

References

Junk & Klar

Succi et al.

See also: Swalbe.∇²f!

Swalbe.∇²f! — Method

∇²f!(output, f, γ)

Finite difference operator for a second derivative in two dimensions.

Computes the laplacian of an input f times a scalar γ and stores the result in output.

Mathematics

The laplacian operator in two dimensions can be written as

$\nabla^2 f = \frac{\partial^2 f}{\partial x^2} + \frac{\partial^2 f}{\partial y}.$

For the discretization of this operator we use a nine point stencil, such the neighbors as well as the diagonal elements. The concrete derivation can be found in the references below, we just show the final result

$\nabla^2 f = \frac{1}{6}\bigg[4(f_{i+1,j} + f_{i,j+1} + f_{i-1,j} + f_{i,j-1}) \newline \qquad\qquad +(f_{i+1,j+1} + f_{i-1,j+1} + f_{i-1,j-1} + f_{i+1,j-1}) \newline \qquad\qquad -20f_{i,j}\bigg] ,$

where we have used Julia conventions, downwards (left) is positive. The whole expression can be multiplied with a scalar γ if needed.

Examples

julia> using Swalbe, Test

julia> arg = reshape(collect(1.0:25),5,5)
5×5 Array{Float64,2}:
 1.0   6.0  11.0  16.0  21.0
 2.0   7.0  12.0  17.0  22.0
 3.0   8.0  13.0  18.0  23.0
 4.0   9.0  14.0  19.0  24.0
 5.0  10.0  15.0  20.0  25.0

julia> res = zeros(5,5); Swalbe.∇²f!(res, arg, -1.0)

julia> analytics = [-30.0 -5.0 -5.0 -5.0 20;
                    -25.0 0.0 0.0 0.0 25.0;
                    -25.0 0.0 0.0 0.0 25.0;
                    -25.0 0.0 0.0 0.0 25.0;
                    -20.0 5.0 5.0 5.0 30.0];

julia> for i in eachindex(analytics)
           @test analytics[i] ≈ res[i] atol=1e-10
       end

References

Junk & Klar

Succi et al.

See also: Swalbe.∇f!